+++ sample/src/main/java/com/squareup/timessquare/sample/SampleTimesSquareActivity.java
@@ -14,7 +14,6 @@ import com.squareup.timessquare.CalendarCellDecorator;
 import com.squareup.timessquare.CalendarPickerView;
 import com.squareup.timessquare.CalendarPickerView.SelectionMode;
 import com.squareup.timessquare.DefaultDayViewAdapter;
-
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Calendar;
@@ -63,6 +62,7 @@ public class SampleTimesSquareActivity extends Activity {
     final Button decorator = (Button) findViewById(R.id.button_decorator);
     final Button hebrew = (Button) findViewById(R.id.button_hebrew);
     final Button arabic = (Button) findViewById(R.id.button_arabic);
+    final Button arabicDigits = (Button) findViewById(R.id.button_arabic_with_digits);
     final Button customView = (Button) findViewById(R.id.button_custom_view);
 
     modeButtons.addAll(Arrays.asList(single, multi, range, displayOnly, decorator, customView));
@@ -106,14 +106,12 @@ public class SampleTimesSquareActivity extends Activity {
 
         calendar.setCustomDayView(new DefaultDayViewAdapter());
         calendar.setDecorators(Collections.<CalendarCellDecorator>emptyList());
-        calendar.init(getDateWithYear(2000), getDateWithYear(2020)) // 20 years, enough to show performance failure
-                .inMode(SelectionMode.SINGLE)
-                .withSelectedDate(c.getTime());
+        // 20 years, enough to show performance failure.
+        calendar.init(getDateWithYear(2000), getDateWithYear(2020))
+            .inMode(SelectionMode.SINGLE).withSelectedDate(c.getTime());
 
         calendar.highlightDates(getHighlightedDaysForMonth( // Adds some highlighted days
-                c.get(Calendar.MONTH) - 1,
-                c.get(Calendar.MONTH),
-                c.get(Calendar.MONTH) + 1));
+            c.get(Calendar.MONTH) - 1, c.get(Calendar.MONTH), c.get(Calendar.MONTH) + 1));
       }
     });
 
@@ -160,8 +158,7 @@ public class SampleTimesSquareActivity extends Activity {
     customized.setOnClickListener(new OnClickListener() {
       @Override public void onClick(View view) {
         showCalendarInDialog("Pimp my calendar!", R.layout.dialog_customized);
-        dialogView.init(lastYear.getTime(), nextYear.getTime())
-            .withSelectedDate(new Date());
+        dialogView.init(lastYear.getTime(), nextYear.getTime()).withSelectedDate(new Date());
       }
     });
 
@@ -181,13 +178,21 @@ public class SampleTimesSquareActivity extends Activity {
       @Override public void onClick(View view) {
         showCalendarInDialog("I'm Hebrew!", R.layout.dialog);
         dialogView.init(lastYear.getTime(), nextYear.getTime(), new Locale("iw", "IL")) //
-                .withSelectedDate(new Date());
+            .withSelectedDate(new Date());
       }
     });
 
     arabic.setOnClickListener(new OnClickListener() {
       @Override public void onClick(View view) {
         showCalendarInDialog("I'm Arabic!", R.layout.dialog);
+        dialogView.init(lastYear.getTime(), nextYear.getTime(), new Locale("ar", "EG")) //
+            .withSelectedDate(new Date());
+      }
+    });
+
+    arabicDigits.setOnClickListener(new OnClickListener() {
+      @Override public void onClick(View view) {
+        showCalendarInDialog("I'm Arabic with Digits!", R.layout.dialog_digits);
         dialogView.init(lastYear.getTime(), nextYear.getTime(), new Locale("ar", "EG")) //
             .withSelectedDate(new Date());
       }
@@ -200,8 +205,8 @@ public class SampleTimesSquareActivity extends Activity {
         calendar.setDecorators(Collections.<CalendarCellDecorator>emptyList());
         calendar.setCustomDayView(new SampleDayViewAdapter());
         calendar.init(lastYear.getTime(), nextYear.getTime())
-                .inMode(SelectionMode.SINGLE)
-                .withSelectedDate(new Date());
+            .inMode(SelectionMode.SINGLE)
+            .withSelectedDate(new Date());
       }
     });